+++ Runtime/Application Manager/Events/ApplicationEvent.cs
@@ -2,27 +2,25 @@
 using UnityEngine;
 
 namespace Application_Manager.Events {
-    [CreateAssetMenu(fileName = "New Application Event", menuName = "Application Event")]
+    [CreateAssetMenu(fileName = "New Application Event",
+        menuName = "Application Manager/Appplication Event/New Application Event")]
     public class ApplicationEvent : ScriptableObject {
-        private readonly List<ApplicationEventListener> eventListeners = 
+        private readonly List<ApplicationEventListener> m_eventListeners =
             new List<ApplicationEventListener>();
 
-        public void Raise()
-        {
-            for(int i = eventListeners.Count -1; i >= 0; i--)
-                eventListeners[i].OnEventRaised();
+        public void Raise() {
+            for (int i = m_eventListeners.Count - 1; i >= 0; i--)
+                m_eventListeners[i].OnEventRaised();
         }
 
-        public void RegisterListener(ApplicationEventListener listener)
-        {
-            if (!eventListeners.Contains(listener))
-                eventListeners.Add(listener);
+        public void RegisterListener(ApplicationEventListener listener) {
+            if (!m_eventListeners.Contains(listener))
+                m_eventListeners.Add(listener);
         }
 
-        public void UnregisterListener(ApplicationEventListener listener)
-        {
-            if (eventListeners.Contains(listener))
-                eventListeners.Remove(listener);
-        } 
+        public void UnregisterListener(ApplicationEventListener listener) {
+            if (m_eventListeners.Contains(listener))
+                m_eventListeners.Remove(listener);
+        }
     }
 }