+++ Runtime/Application Manager/States/BaseState.cs
@@ -2,6 +2,8 @@
 
 namespace Application_Manager.States {
     public abstract class BaseState : ScriptableObject, IState {
+        public const string menuRootName = "Application Manager/States/";
+        
         public virtual void OnEnter() {
             // Noop
         }