+++ Runtime/Application Manager/States/InitialState.cs
@@ -2,8 +2,7 @@
 using UnityEngine;
 
 namespace Application_Manager.States {
-    // TODO: All Create Asset Menu should be in the menu name "Application Manager/..."
-    [CreateAssetMenu(fileName = "New Initial State", menuName = "States/Initial State")]
+    [CreateAssetMenu(fileName = "New Initial State", menuName = menuRootName + "Initial State")]
     public class InitialState : BaseState {
         public ApplicationEventListener eventListener;