+++ CHANGELOG.md
@@ -1,5 +1,13 @@
+## [0.0.2] - 14/03/2025
+
+### Updated Submenu to create states, events and listeners
+
+- Added submenu to create states, events and listeners all from the same parent
+
 ## [0.0.1] - 14/03/2025
+
 ### Initial Release
+
 - Added Improved Game Manager
 - Can create states using scriptable objects
 - Can change states using scriptable objects events